Moved snippets test models into separate app
