+++ wagtail/tests/snippets/__init__.py
@@ -1,0 +1,1 @@
+default_app_config = 'wagtail.tests.snippets.apps.WagtailSnippetsTestsAppConfig'
