+++ wagtail/tests/snippets/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+
+
+class WagtailSnippetsTestsAppConfig(AppConfig):
+    name = 'wagtail.tests.snippets'
+    label = 'snippetstests'
+    verbose_name = "Wagtail snippets tests"
